backend/requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt
python-jose
google-genai
python-dotenv
email-validator

